Better naming `ActiveRecord` for `Model`.
