+++ composer.json
@@ -64,16 +64,18 @@
         },
         "config-plugin-environments": {
             "tests": {
-                "console": "tests/console/*.php",
+                "common": "tests/common/*.php",
+                "console": [
+                    "$common",
+                    "tests/console/*.php"
+                ],
                 "web": [
+                    "$common",
                     "web/*.php",
                     "tests/web/*.php"
                 ],
                 "params": "params.php",
-                "params-console": [
-                    "$params",
-                    "tests/params-console.php"
-                ],
+                "params-console": "$params",
                 "params-web": [
                     "$params",
                     "tests/params-web.php"
