--- config/tests/params-console.php
@@ -1,5 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-return [];
