+++ src/ActiveRecord/Token.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yii\User\Model;
+namespace Yii\User\ActiveRecord;
 
 use yii\db\ActiveQueryInterface;
 use yii\db\ActiveRecord;
@@ -11,23 +11,24 @@ use yii\db\ActiveRecord;
  * Represents the data structure for a token object.
  *
  * @property int $id
- * @property string $provider
- * @property string $client_id
- * @property string $data
  * @property string $code
  * @property int $created_at
- * @property string $email
- * @property string $username
+ * @property int $type
  */
-final class SocialAccount extends ActiveRecord
+final class Token extends ActiveRecord
 {
     public function getIdentity(): ActiveQueryInterface
     {
         return $this->hasOne(Identity::class, ['id' => 'id']);
     }
 
+    public static function primaryKey(): array
+    {
+        return ['id', 'code', 'type'];
+    }
+
     public static function tableName(): string
     {
-        return '{{%social_account}}';
+        return '{{%token}}';
     }
 }
