+++ src/Framework/Migration/M211126112534Identity.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use Yii\User\ActiveRecord\Identity;
 use Yii\User\Framework\Migration\Migration;
 
 final class M211126112534Identity extends Migration
@@ -9,7 +10,7 @@ final class M211126112534Identity extends Migration
     public function up(): bool
     {
         $this->createTable(
-            '{{%identity}}',
+            Identity::tableName(),
             [
                 'id' => $this->primaryKey()->notNull()->unsigned(),
                 'auth_key' => $this->string(32)->defaultValue(''),
@@ -22,7 +23,7 @@ final class M211126112534Identity extends Migration
 
     public function down(): bool
     {
-        $this->dropTable('{{%identity}}');
+        $this->dropTable(Identity::tableName());
 
         return true;
     }
