+++ tests/Unit/RegisterServiceTest.php
@@ -9,12 +9,12 @@ use Codeception\Test\Unit;
 use RuntimeException;
 use Yii;
 use yii\db\ActiveRecordInterface;
+use Yii\User\ActiveRecord\Account;
+use Yii\User\ActiveRecord\Identity;
+use Yii\User\ActiveRecord\Profile;
+use Yii\User\ActiveRecord\SocialAccount;
 use Yii\User\Framework\Repository\PersistenceRepository;
 use Yii\User\Framework\Repository\PersistenceRepositoryInterface;
-use Yii\User\Model\Account;
-use Yii\User\Model\Identity;
-use Yii\User\Model\Profile;
-use Yii\User\Model\SocialAccount;
 use Yii\User\UseCase\Register\RegisterForm;
 use Yii\User\UseCase\Register\RegisterService;
 use Yii\User\UserModule;
